refactor: PokemonUiState 를 도입한다
