+++ android/app/src/main/java/poke/rogue/helper/presentation/dex/detail/PokemonDetailUiState.kt
@@ -1,0 +1,28 @@
+package poke.rogue.helper.presentation.dex.detail
+
+import poke.rogue.helper.data.model.PokemonDetail
+import poke.rogue.helper.data.model.Stat
+import poke.rogue.helper.presentation.dex.model.PokemonUiModel
+import poke.rogue.helper.presentation.dex.model.StatUiModel
+import poke.rogue.helper.presentation.dex.model.toUi
+
+sealed interface PokemonDetailUiState {
+    data class PokemonDetailUiModel(
+        val pokemon: PokemonUiModel,
+        val stats: List<StatUiModel>,
+        val abilities: List<String>,
+        val height: Float,
+        val weight: Float,
+    ) : PokemonDetailUiState
+
+    data object IsLoading : PokemonDetailUiState
+}
+
+fun PokemonDetail.toUi(): PokemonDetailUiState.PokemonDetailUiModel =
+    PokemonDetailUiState.PokemonDetailUiModel(
+        pokemon = pokemon.toUi(),
+        stats = stats.map(Stat::toUi),
+        abilities = abilities,
+        height = height,
+        weight = weight,
+    )
